fs: boot: Update bootmeth_alloc_other() to use abuf
Using an abuf for this function simplifies returning the size and also
makes it easier to free memory afterwards. Update the API and callers.

+++ include/bootflow.h
@@ -101,7 +101,7 @@ struct bootflow {
 	char *subdir;
 	char *fname;
 	void *logo;
-	uint logo_size;
+	size_t logo_size;
 	char *buf;
 	int size;
 	int err;
